Only moderators and the reaction author can now delete reactions from the database, merged pr with master

+++ anime.py
@@ -32,6 +32,13 @@ class anime():
     def __init__(self, bot):
         self.bot = bot
 
+    def checkRole(self, user, roleRec):
+            ok = False
+            for all in list(user.roles):
+                if all.name == roleRec:
+                    ok = True
+    return ok
+
     @commands.command(pass_context=True)
     async def kawaii(self, ctx):
         '''Gibt ein zufälliges kawaii Bild aus'''
@@ -82,9 +89,14 @@ class anime():
                     con.commit()
                     await self.bot.say(':ok: Command **{}** hinzugefügt!'.format(arg[0].lower()))
             elif command == 'del':
-                c.execute('DELETE FROM "reactions" WHERE "id" in (?)', (int(arg[0]), ))
-                con.commit()
-                await self.bot.say(':put_litter_in_its_place: ID #{} gelöscht!'.format(arg[0].lower()))
+                if self.checkRole(ctx.message.author, self.mod):
+                    c.execute('DELETE FROM "reactions" WHERE "id" in (?)', (int(arg[0]), ))
+                    con.commit()
+                    await self.bot.say(':put_litter_in_its_place: ID #{} gelöscht!'.format(arg[0].lower()))
+                else:
+                    c.execute('DELETE FROM "reactions" WHERE "id" in (?) AND "author" IN (?)', (int(arg[0]), str(ctx.message.author)))
+                    con.commit()
+                    await self.bot.say(':put_litter_in_its_place: ID #{} gelöscht!'.format(arg[0].lower()))
             elif command == 'list':
                 lst = c.execute('SELECT * FROM "reactions"')
                 msg = ''
